pyproject.toml
[tool.poetry]
name = "goit-pythonweb-hw-10"
version = "0.1.0"
description = ""
authors = ["Anton Povar <povar.anton@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10"
fastapi = ">=0.115.12,<0.116.0"
sqlalchemy = ">=2.0.40,<3.0.0"
asyncpg = ">=0.30.0,<0.31.0"
alembic = ">=1.15.2,<2.0.0"
pydantic = {extras = ["email"], version = ">=2.11.1,<3.0.0"}
faker = ">=37.1.0,<38.0.0"
pydantic-settings = ">=2.8.1,<3.0.0"
uvicorn = {extras = ["standard"], version = ">=0.34.0,<0.35.0"}
libgravatar = "^1.0.4"
python-jose = {extras = ["cryptography"], version = "^3.4.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.20"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
